Do not allow shared+static
I recently wanted to compile a project against static libs and realized that
there was no easy way for me to add the required __TBB_LIB_NAME=tbb_static to select the right library name externally.

Perhaps it's better to not allow for shared & static to be built at the same time?